Moved things around a little
