+++ src/common.css
@@ -1,0 +1,4 @@
+.fullSize {
+  height: 100%;
+  overflow: hidden;
+}
